hacks(*) avoid java 15 to take over
mb doesn't work in java 15. if we run :shell with bash -l, it loads
everything in /etc/profile.d, and there is 00-resore-env.sh, that sets
java 15 in the path.

We play a huge hack by overwritting our profile.sh into that file,
clobbering its effect and leaving us with a java 8.